Remove 404 GHCR link from README
I happened to come across this repository during work and was reading the README when I noticed that this URL returns a 404 Not Found error.

Since I am not familiar with the historical context of this project, I am not entirely sure, but does this mean that the GHCR (GitHub Container Registry) version is no longer maintained? If so, I thought the link should be removed.